+++ piet-gpu/src/simple_text.rs
@@ -22,7 +22,7 @@ pub use pinot::FontRef;
 
 // This is very much a hack to get things working.
 // On Windows, can set this to "c:\\Windows\\Fonts\\seguiemj.ttf" to get color emoji
-const FONT_DATA: &[u8] = include_bytes!("../third-party/Roboto-Regular.ttf");
+const FONT_DATA: &[u8] = include_bytes!("../../piet-wgsl/examples/assets/third-party/Roboto-Regular.ttf");
 
 pub struct SimpleText {
     gcx: GlyphContext,
