+++ piet-wgsl/examples/winit/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "winit"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+piet-wgsl = { path = "../../../piet-wgsl" }
+piet-scene = { path = "../../../piet-scene" }
+winit = "0.27.5"
+pollster = "0.2.5"
+# for picosvg
+roxmltree = "0.13"
