+++ piet-wgsl/examples/winit/src/pico_svg.rs
