--- piet-wgsl/src/debug.rs
@@ -1,5 +1,0 @@
-#![allow(dead_code)]
-
-pub mod clip;
-pub mod draw;
-pub mod fine;
